requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
bandit==1.7.8
safety==3.2.5
pre-commit==3.4.0
pytest==8.3.3
coverage==7.6.1
